% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLOddsExtractor.R
\name{geneLOddsExtractor}
\alias{geneLOddsExtractor}
\title{geneLOddsExtractor}
\usage{
geneLOddsExtractor(scObj, hgnc = T, intercept = T, na.rm = T,
  loddsum = F)
}
\arguments{
\item{scObj}{A sparseClassification object as obtained from the macClassifier function.}

\item{hgnc}{A boolean argument whether you want the log odds with Entrez (FALSE) or HGNC symbols (TRUE). Defaults to TRUE.}

\item{intercept}{A boolean argument whether you want the intercept to be returned as well. Defaults to TRUE.}

\item{na.rm}{A boolean argument whether you want to remove absent classifier genes. Defaults to TRUE.}

\item{loddsum}{A boolean argument on whether you want to sum the log odds per activation state. Defaults to FALSE.}
}
\description{
This function extracts the gene log odds and stores it in a dataframe
}
