% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLOddsPlot.R
\name{geneLOddsPlot}
\alias{geneLOddsPlot}
\title{geneLOddsPlot}
\usage{
geneLOddsPlot(scObj, samples = NULL, phenotype = NULL, title = NULL,
  hgnc = T)
}
\arguments{
\item{scObj}{A sparseClassification object as obtained from the macClassifier function.}

\item{samples}{(optional) A vector including the samples to be extracted. Uses the column names of the expression data.}

\item{phenotype}{(optional) A vector of length equal to the number of samples included in the classification containing the actual classes.
Used for coloring the bars.}

\item{title}{(optional) A title for the plot.}

\item{hgnc}{A boolean argument whether you want the log odds with Entrez or HGNC symbols. Defaults to HGNC symbols.}
}
\description{
This function visualizes the log odds of the individual genes.
}
