% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lORStabilizer.R
\name{lORStabilizer}
\alias{lORStabilizer}
\title{lORStabilizer}
\usage{
lORStabilizer(threshold = 0.5, aggregation_method = c("median",
  "mean"))
}
\arguments{
\item{threshold}{The proportion of the iterations that must be non-zero for a gene to be considered stable.}

\item{aggregation_method}{Which method for aggregation across the iterations should be used? Options are taking the median ("median")
or taking the mean ("mean"). Defaults to "median".}
}
\description{
This function stabilizes the coefficients as obtained from the repeated cross-validation data ("repcv_lor").
}
