% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macClassifier.R
\name{macClassifier}
\alias{macClassifier}
\title{macClassifier}
\usage{
macClassifier(exprs_data, lor_list, intercept = T)
}
\arguments{
\item{exprs_data}{A matrix with rows representing genes and columns representing samples.}

\item{lor_list}{A list of log odds ratios per macrophage subset.}

\item{intercept}{A boolean whether to include the intercept}
}
\description{
This function takes the expression data in matrix format and the stabilized log odds ratios (as calculated using the lORStabilizer
function) and calculates the response for different macrophage subsets. #' Currently, the expression data can be classified as
"CTRL", "LPS_Early", "LPS_Late", "LPS_IFNg", "IFNg", "IL4", "IL10", or "DEX" macrophage. The results are then stored in a
sparseClassification object, which can be queried using specific accessor methods.
}
\examples{

#Prepare the test data
data(testset_exprdata)
head(testset_exprdata)

#Prepare the stabilized odds ratios (defaults to 0.5 median stabilization)
stab_lor_list <- lORStabilizer()

#Calculate the predictions for the test data per class
testset_spobj <- macClassifier(testset_exprdata, stab_lor_list)

#Extract the predictions [0,1] (highest value is the predicted class)
testset_props <- getProportion(testset_spobj)

}
