% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsePlot.R
\name{responsePlot}
\alias{responsePlot}
\title{responsePlot}
\usage{
responsePlot(scObj, actual_class = NULL,
  type_response = c("proportion", "odds", "lodds"))
}
\arguments{
\item{scObj}{A sparseClassification object as obtained from the macClassifier function.}

\item{actual_class}{An optional vector of length equal to the number of samples included in the classification containing the actual classes.
Used for coloring the bars.}

\item{type_response}{Should the response be a proportion ("proportion"; [0;1]), odds ("odds"; [-Inf;Inf]), log odds ("lodds"; [0;Inf]),
or should all aforementioned values be returned as separate list elements ("all"). Defaults to "proportion".}
}
\description{
This function acts as a wrapper to plot the responses.
}
