% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseOdds.R
\name{sparseOdds}
\alias{sparseOdds}
\title{sparseOdds}
\usage{
sparseOdds(stab_lor_list, exprs_data, expon = TRUE)
}
\arguments{
\item{stab_lor_list}{A list of odds ratios annotated with Entrez IDs. Can simply use the output of the coef_stabilizer() function.}

\item{exprs_data}{A matrix with rows representing genes and columns representing samples.}

\item{expon}{A boolean argument whether or not to return the odds (TRUE) or the log odds (FALSE). Defaults to TRUE.}
}
\description{
This function calculates the (log) odds for the individual genes composing each classifier.
}
